app/requirements.txt
simple-log-factory == 0.0.1
raccoon-simple-stopwatch == 0.0.3
requests == 2.30.0
speedtest-cli == 2.1.3
paho-mqtt == 1.6.1